requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
matplotlib>=3.8
PyYAML>=6.0
lightgbm>=4.3
scikit-learn>=1.4
scipy>=1.11
joblib>=1.3
google-cloud-storage>=2.16
google-cloud-firestore>=2.16
